Fix sensor activate after SensorEventConnection removed
When app idle and register a sensor on existing connection,the method
sensor.batch() retrun NO_ERROR but sensor.activate() return INVALID_OPERATION.
The sensor will remove from the connecion but unable to remove from the
batchParams,this make the sensor active after the app killed.
As the client is disabled, keep the sensor in the connection for resume.

Bug: 180825144
Test: Debug app in 180825144

Signed-off-by: xiamengsen <xiamengsen@xiaomi.com>
Change-Id: Icc22862d1a083abba08fb2cbcabd043444b7090f
Signed-off-by: anulalat <ndra.bodoh@gmail.com>
Signed-off-by: HeroBuxx <herobuxx@conqueros.co>